Refactor finetune module and restore blueprint architecture
- Restore app.py to main's blueprint pattern (~55 lines) and register
  finetune_bp alongside existing route blueprints
- Extract finetune helpers into dashboard/finetune_utils.py
- Create dashboard/routes/finetune_routes.py as a Flask Blueprint
- Add finetune state vars to dashboard/state.py
- Rename finetune files for clarity: dataset.py -> correction_dataset.py,
  trainer.py -> lora_trainer.py, cli.py -> finetune_cli.py,
  job_manager.py -> finetune_job_manager.py,
  model_templates.py -> finetuned_model_templates.py
- Remove debug prints and trim verbose docstrings
- Use ImageDataInterface for correction data loading
