+++ cellmap_flow/finetune/finetuned_model_templates.py
